test(auth): 토큰 발급 기능(로그인) 인수 테스트 만들기

+++ src/main/java/nextstep/subway/path/domain/PathFinder.java
@@ -44,8 +44,8 @@ public class PathFinder {
 
     private Path createPath(Station source, Station target) {
         List<Station> stations = find.getVertexList(source, target);
-        double weight = find.getWeight(source, target);
-        return new Path(stations, (int) weight);
+        int weight = find.getWeight(source, target);
+        return new Path(stations, weight);
     }
 
     private void addSectionsInGraph(List<Line> lines) {
